# Optimize MessageCracker.IsHandlerMethod for memory